requirements.txt
pony~=0.7.17
falcon-auth==1.1.0 
PyJWT==2.4.0
falcon-multipart==0.2.0 
PyMySQL==1.0.2
requests==2.26.0 
oauthlib==3.1.1
requests-oauthlib==1.3.0
falcon-swagger-ui==1.2.2
gunicorn==20.1.0
elastic-apm
flask_socketio
pytz
elasticsearch==7.15.2
boto3
pdfkit
Jinja2
Pillow
PyPDF2
memory-profiler
mysql.connector
coverage
openpyxl
user_agents==2.2.0
google-auth
cryptography
bleach==6.0.0
# bleach[css]
tinycss2
psutil~=5.9.7
falcon~=3.1.3
lxml~=4.9.4
Scrapy~=2.11.0